Add missing annotations, type variables and some minor cleanup.
git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@466886 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/awt/src/main/java/common/org/apache/harmony/awt/gl/render/NativeImageBlitter.java
@@ -64,8 +64,8 @@ public class NativeImageBlitter implements Blitter {
             }else{
                 double scaleX = xform.getScaleX();
                 double scaleY = xform.getScaleY();
-                double scaledX = (double)dstX / scaleX;
-                double scaledY = (double)dstY / scaleY;
+                double scaledX = dstX / scaleX;
+                double scaledY = dstY / scaleY;
                 AffineTransform at = new AffineTransform();
                 at.setToTranslation(scaledX, scaledY);
                 xform.concatenate(at);
